Separate vulkan impl from window.cpp
